# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Suivi VtES API",
    "description": "API du suivi de pratique VtES (collection EN/FR, decks, parties, tournois). Pilote d'une PWA offline-first destinée à être portée sur barrins-project.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/decks": {
      "get": {
        "tags": [
          "decks"
        ],
        "summary": "Liste les decks",
        "operationId": "listDecks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DeckStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeckRead"
                  },
                  "title": "Response Listdecks"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "decks"
        ],
        "summary": "Crée un deck",
        "operationId": "createDeck",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckRead"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/decks/{deck_id}": {
      "get": {
        "tags": [
          "decks"
        ],
        "summary": "Un deck et sa composition",
        "operationId": "getDeck",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckDetailRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "decks"
        ],
        "summary": "Modifie un deck",
        "description": "Passer un deck à `active` exige qu'il soit légal (crypt ≥ 12, library 60–90) ; sinon 409.",
        "operationId": "updateDeck",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "decks"
        ],
        "summary": "Supprime un deck et sa composition",
        "operationId": "deleteDeck",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/decks/{deck_id}/legalite": {
      "get": {
        "tags": [
          "decks"
        ],
        "summary": "Légalité du deck",
        "description": "Calculée à la demande ; les seuils voyagent dans la réponse.",
        "operationId": "getDeckLegality",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckLegality"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/decks/{deck_id}/cartes": {
      "post": {
        "tags": [
          "decks"
        ],
        "summary": "Ajoute une carte au deck",
        "description": "La carte doit être en collection dans la langue demandée, avec assez d'exemplaires disponibles (hors proxies) ; sinon 409.",
        "operationId": "addDeckCard",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckCardCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckCardRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/decks/{deck_id}/cartes/{card_id}/{language_code}": {
      "patch": {
        "tags": [
          "decks"
        ],
        "summary": "Modifie une ligne du deck",
        "operationId": "updateDeckCard",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          },
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "language_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeckCardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeckCardRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "decks"
        ],
        "summary": "Retire une carte du deck",
        "operationId": "removeDeckCard",
        "parameters": [
          {
            "name": "deck_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Deck Id"
            }
          },
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          },
          {
            "name": "language_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Language Code"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CardCategory": {
        "type": "string",
        "enum": [
          "crypt",
          "library"
        ],
        "title": "CardCategory",
        "description": "Les deux moitiés d'un deck VtES (CLAUDE.md §5)."
      },
...
      "CardSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vekn_id": {
            "type": "integer",
            "title": "Vekn Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/CardCategory"
          },
          "clan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClanRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "group_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Code"
          },
          "advanced": {
            "type": "boolean",
            "title": "Advanced",
            "default": false
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Scan de la carte (version anglaise de référence)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "vekn_id",
          "name",
          "category"
        ],
        "title": "CardSummary",
        "description": "Vue courte d'une carte, pour les listes et l'autocomplétion."
      },
...
      "ClanRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "abbrev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbrev"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ClanRead",
        "description": "Clan."
      },
...
      "DeckCardCreate": {
        "properties": {
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity"
          },
          "proxy_quantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Proxy Quantity",
            "default": 0
          },
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "language_code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Language Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "quantity",
          "card_id",
          "language_code"
        ],
        "title": "DeckCardCreate",
        "description": "Ajout d'une carte à un deck.\n\nLa carte doit déjà exister dans la collection pour la langue demandée\n(CLAUDE.md §11 point 2) ; la vérification de disponibilité relève du\nservice, la base garantissant déjà l'existence de l'entrée de collection."
      },
      "DeckCardRead": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "language_code": {
            "type": "string",
            "title": "Language Code"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "proxy_quantity": {
            "type": "integer",
            "title": "Proxy Quantity",
            "description": "Part des exemplaires ci-dessus jouée en proxy."
          },
          "card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "language_code",
          "quantity",
          "proxy_quantity"
        ],
        "title": "DeckCardRead",
        "description": "Une ligne de decklist."
      },
      "DeckCardUpdate": {
        "properties": {
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity"
          },
          "proxy_quantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Proxy Quantity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DeckCardUpdate",
        "description": "Modification d'une ligne de decklist.\n\nLe contrôle `proxy_quantity <= quantity` n'est complet que si les deux\nvaleurs sont fournies. Sinon la comparaison porterait sur une valeur que\nseule la base connaît : **le service du Lot 2 doit refaire la vérification**\naprès fusion avec la ligne existante (la contrainte `CHECK` de `deck_card`\nreste le dernier filet, mais elle produirait un 500 plutôt qu'un 422)."
      },
      "DeckCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "created_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created On"
          },
          "status": {
            "$ref": "#/components/schemas/DeckStatus",
            "default": "draft"
          },
          "archetype": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DeckCreate",
        "description": "Création d'un deck."
      },
      "DeckDetailRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created On"
          },
          "status": {
            "$ref": "#/components/schemas/DeckStatus"
          },
          "archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/DeckCardRead"
            },
            "type": "array",
            "title": "Cards",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DeckDetailRead",
        "description": "Un deck avec sa composition."
      },
      "DeckLegality": {
        "properties": {
          "deck_id": {
            "type": "integer",
            "title": "Deck Id"
          },
          "crypt_count": {
            "type": "integer",
            "title": "Crypt Count"
          },
          "library_count": {
            "type": "integer",
            "title": "Library Count"
          },
          "crypt_minimum": {
            "type": "integer",
            "title": "Crypt Minimum"
          },
          "library_minimum": {
            "type": "integer",
            "title": "Library Minimum"
          },
          "library_maximum": {
            "type": "integer",
            "title": "Library Maximum"
          },
          "is_legal": {
            "type": "boolean",
            "title": "Is Legal"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "deck_id",
          "crypt_count",
          "library_count",
          "crypt_minimum",
          "library_minimum",
          "library_maximum",
          "is_legal"
        ],
        "title": "DeckLegality",
        "description": "Verdict de légalité d'un deck (CLAUDE.md §5 : crypt ≥ 12, library 60–90).\n\nSchéma de sortie seulement : le calcul est une règle de service, écrite au\nLot 2 avec la skill `regles-vtes`. Les seuils voyagent dans la réponse pour\nque le front affiche « 58 / 60 » sans les redéfinir de son côté."
      },
      "DeckRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created On"
          },
          "status": {
            "$ref": "#/components/schemas/DeckStatus"
          },
          "archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "DeckRead",
        "description": "Un deck, sans sa composition."
      },
      "DeckStatus": {
        "type": "string",
        "enum": [
          "draft",
          "active",
          "retired"
        ],
        "title": "DeckStatus",
        "description": "Cycle de vie d'un deck côté joueur."
      },
      "DeckUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "created_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created On"
          },
          "status": {
            "$ref": "#/components/schemas/DeckStatus"
          },
          "archetype": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Archetype"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DeckUpdate",
        "description": "Modification partielle d'un deck.\n\n`name` et `status` sont facultatifs mais non nullables ; `created_on`,\n`archetype` et `notes` acceptent `null` pour effacer la valeur."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Explication lisible de l'échec."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Corps d'une réponse d'erreur métier."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}